Added `ser/tuple_structs` module
